# swagger.yaml
...
openapi: 3.0.0
info:
  title: API DE 31031669 - Tarea 1
  description: |
    API RESTful con autenticación JWT y gestión de usuarios
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://p3-31031669-2.onrender.com
    description: Servidor de producción
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Obtener todos los usuarios
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              example:
                status: "success"
                data:
                  - id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    fullName: "Adriana Padron"
                    email: "adri@example.com"
                    cedula: "12345678"
                    createdAt: "2024-01-01T00:00:00.000Z"
                    updatedAt: "2024-01-01T00:00:00.000Z"
                  - id: "b2c3d4e5-f6g7-8901-bcde-f23456789012"
                    fullName: "Adriana Padron"
                    email: "adri@example.com"
                    cedula: "87654321"
                    createdAt: "2024-01-01T00:00:00.000Z"
                    updatedAt: "2024-01-01T00:00:00.000Z"

    post:
      tags:
        - users
      summary: Crear nuevo usuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - email
                - password
                - cedula
              properties:
                fullName:
                  type: string
                  example: "Adriana Padron"
                email:
                  type: string
                  example: "adri@example.com"
                password:
                  type: string
                  example: "password123"
                cedula:
                  type: string
                  example: "11223344"
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              example:
                status: "success"
                data:
                  id: "c3d4e5f6-g7h8-9012-cdef-345678901234"
                  fullName: "Adriana Padron"
                  email: "adri@example.com"
                  cedula: "11223344"
                  createdAt: "2024-01-01T00:00:00.000Z"
                  updatedAt: "2024-01-01T00:00:00.000Z"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT